+++ example/lib/main.dart
@@ -121,6 +121,10 @@ class _MyHomePageState extends State<MyHomePage> {
                     radius: 8,
                     hintText: "Phone Number",
                     borderWidth: 2,
+                    backgroundColor: Colors.transparent,
+                    decoration: null,
+                    autoFocus: true,
+                    showCursor: false,
                     textStyle: const TextStyle(
                         color: Colors.black,
                         fontSize: 16,