+++ lib/models/phone_config.dart
@@ -4,13 +4,22 @@ class PhoneConfig {
   Color focusedColor, enabledColor;
   TextStyle textStyle, hintStyle;
   double radius, borderWidth;
+  Decoration? decoration;
+  Color? backgroundColor;
   String? hintText;
+  bool autoFocus, showCursor;
+  FocusNode? focusNode;
   PhoneConfig(
       {this.focusedColor = const Color(0xFF6D59BD),
       this.enabledColor = const Color(0xFF6D59BD),
       this.radius = 8,
       this.hintText = "Phone Number",
       this.borderWidth = 2,
+      this.backgroundColor,
+      this.decoration,
+      this.focusNode,
+      this.autoFocus = false,
+      this.showCursor = false,
       TextStyle? textStyle,
       TextStyle? hintStyle})
       : textStyle = textStyle ??